Changed BrokerageId from int to long
IOrderProvider.GetOrderByBrokerageId now accepts a long value (as required by Oanda and Tradier brokerages)

+++ Tests/Brokerages/OrderProvider.cs
@@ -56,7 +56,7 @@ namespace QuantConnect.Tests.Brokerages
             return _orders.FirstOrDefault(x => x.Id == orderId);
         }
 
-        public Order GetOrderByBrokerageId(int brokerageId)
+        public Order GetOrderByBrokerageId(long brokerageId)
         {
             return _orders.FirstOrDefault(x => x.BrokerId.Contains(brokerageId));
         }
